masking and refactor
