+++ src/math/mod.rs
@@ -7,4 +7,3 @@ pub mod composition;
 pub mod domain;
 pub mod fri;
 pub mod polynomial;
-pub mod stark;
